Notify clients when meals are clarified
