+++ FoodBot/Services/Workers/TextMealQueueWorker.cs
@@ -86,7 +86,7 @@ public sealed class TextMealQueueWorker : BackgroundService
                         db.PendingClarifies.Remove(nextClar);
                         await db.SaveChangesAsync(stoppingToken);
 
-                        await notifier.MealUpdated(meal.ChatId, ToListItem(meal));
+                        await notifier.MealUpdated(meal.ChatId, meal.ToListItem());
                     }
                     catch (Exception ex)
                     {
@@ -139,7 +139,7 @@ public sealed class TextMealQueueWorker : BackgroundService
                     db.PendingMeals.Remove(next);
                     await db.SaveChangesAsync(stoppingToken);
 
-                    await notifier.MealUpdated(entry.ChatId, ToListItem(entry));
+                    await notifier.MealUpdated(entry.ChatId, entry.ToListItem());
                 }
                 catch (DbUpdateConcurrencyException ex)
                 {
@@ -169,24 +169,4 @@ public sealed class TextMealQueueWorker : BackgroundService
         }
     }
 
-    private static MealListItem ToListItem(MealEntry meal)
-    {
-        var ingredients = string.IsNullOrWhiteSpace(meal.IngredientsJson)
-            ? Array.Empty<string>()
-            : (System.Text.Json.JsonSerializer.Deserialize<string[]>(meal.IngredientsJson!) ?? Array.Empty<string>());
-        return new MealListItem(
-            meal.Id,
-            meal.CreatedAtUtc,
-            meal.DishName,
-            meal.WeightG,
-            meal.CaloriesKcal,
-            meal.ProteinsG,
-            meal.FatsG,
-            meal.CarbsG,
-            ingredients,
-            ProductJsonHelper.DeserializeProducts(meal.ProductsJson),
-            meal.ImageBytes != null && meal.ImageBytes.Length > 0,
-            false
-        );
-    }
 }
